+++ ModEntry.cs
@@ -3,20 +3,17 @@ using Microsoft.Xna.Framework;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewValley;
-using System.Reflection;
-using StardewValley.Menus;
 using StardewValley.Monsters;
 using StardewValley.Locations;
-using StardewValley.Characters;
 using System.Collections.Generic;
 using System.Timers;
 using System.Linq;
 using HarmonyLib;
 using Microsoft.Xna.Framework.Graphics;
-using StardewValley.TerrainFeatures;
 using StardewValley.Tools;
 using Newtonsoft.Json;
-using StardewValley.Enchantments;
+using StardewValley.Menus;
+using StardewValley.TerrainFeatures;
 
 namespace LevelExtender
 {
@@ -59,9 +56,7 @@ namespace LevelExtender
 
         public List<Skill> skills = new List<Skill>();
         public static List<int[]> categories = new List<int[]>();
-
-        JsonSerializer serializer = new JsonSerializer();
-
+        
         public static List<int> skillLevs = new List<int>();
 
         public ModEntry()
@@ -73,9 +68,6 @@ namespace LevelExtender
             s_mod = -1.0;
             mpload = false;
             mpMult = 1.0;
-
-            serializer.NullValueHandling = NullValueHandling.Ignore;
-            serializer.Formatting = Formatting.Indented;
         }
 
         public override object GetApi()
@@ -114,7 +106,7 @@ namespace LevelExtender
             helper.Events.Display.Rendered += this.Display_Rendered;
             helper.Events.Player.Warped += this.Player_Warped;
             helper.Events.Content.AssetRequested += this.OnAssetRequested;
-            helper.Events.World.NpcListChanged += this.OnNpcListChanged; // <-- ADD THIS EVENT
+            helper.Events.World.NpcListChanged += this.OnNpcListChanged;
 
             helper.ConsoleCommands.Add("xp", "Displays the xp table for your current skill levels.", this.XPT);
             helper.ConsoleCommands.Add("lev", "Sets the player's level: lev <skill name> <number>", this.SetLev);
@@ -130,7 +122,6 @@ namespace LevelExtender
             LEE.OnXPChanged += this.OnXPChanged;
         }
         
-        // ADD THIS NEW METHOD
         private void OnNpcListChanged(object sender, NpcListChangedEventArgs e)
         {
             if (!e.IsCurrentLocation || !e.Location.IsFarm)
@@ -146,13 +137,12 @@ namespace LevelExtender
             }
         }
 
-       private void OnAssetRequested(object sender, AssetRequestedEventArgs e)
+        private void OnAssetRequested(object sender, AssetRequestedEventArgs e)
         {
             if (e.Name.IsEquivalentTo("Data/Fish"))
             {
                 e.Edit(asset =>
                 {
-                    // The dictionary now uses <string, string> instead of <int, string>
                     IDictionary<string, string> data = asset.AsDictionary<string, string>().Data;
                     foreach (var pair in data.ToArray())
                     {
@@ -612,7 +602,7 @@ namespace LevelExtender
                 m.Speed += rand.Next((int)Math.Round((Game1.player.combatLevel.Value / 10.0)));
                 m.resilience.Value = m.resilience.Value + (Game1.player.combatLevel.Value / 10);
                 m.ExperienceGained += (int)(m.Health / 100.0) + ((10 + (Game1.player.combatLevel.Value * 2)) * tier);
-                
+
                 Game1.currentLocation.characters.Add((NPC)m);
                 total_m++;
 
@@ -913,198 +903,6 @@ namespace LevelExtender
             if (sk == null) return -1;
             skills.Add(sk);
             return 0;
-        }
-    }
-
-    public class LEEvents
-    {
-        public event EventHandler<EXPEventArgs> OnXPChanged;
-        public void RaiseEvent(EXPEventArgs args)
-        {
-            OnXPChanged?.Invoke(this, args);
-        }
-    }
-
-    public class EXPEventArgs : EventArgs
-    {
-        public int key { get; set; }
-    }
-
-    public class XPBar
-    {
-        public Skill skill;
-        public Timer timer;
-        public DateTime time;
-        public double ych = 0;
-        public bool startmove = false;
-        public int movedir = 0;
-
-        public XPBar(Skill skill)
-        {
-            this.skill = skill;
-            this.timer = new Timer(5000);
-            timer.Elapsed += delegate { skill.LE.EndXPBar(skill.key); };
-            timer.AutoReset = false;
-            timer.Enabled = true;
-            time = DateTime.Now;
-        }
-
-        public int ychi
-        {
-            get => (int)Math.Round(ych);
-            set { } // This setter was incorrect, leaving it empty is safer than the recursive call it had.
-        }
-    }
-
-    public class Skill
-    {
-        public string name;
-        public int key;
-        public bool lvlbyxp = false;
-        private int Level;
-        public int level
-        {
-            get { return Level; }
-            set
-            {
-                Level = value;
-                if (key < 5)
-                {
-                    if (key == 0) Game1.player.farmingLevel.Value = level;
-                    else if (key == 1) Game1.player.fishingLevel.Value = level;
-                    else if (key == 2) Game1.player.foragingLevel.Value = level;
-                    else if (key == 3) Game1.player.miningLevel.Value = level;
-                    else if (key == 4) Game1.player.combatLevel.Value = level;
-                }
-                if (!lvlbyxp)
-                {
-                    int reqxp = getReqXP(level - 1);
-                    if (key < 5)
-                        Game1.player.experiencePoints[key] = reqxp;
-                    XP = reqxp;
-                }
-                lvlbyxp = false;
-            }
-        }
-        public int xpc;
-        public EXPEventArgs args;
-        public ModEntry LE;
-        public List<int> xp_table;
-        public double xp_mod;
-        private int XP;
-        public int xp
-        {
-            get { return XP; }
-            set
-            {
-                if (xp != value)
-                {
-                    xpc = value - xp;
-                    XP = value;
-                    checkForLevelUp();
-                    LE.LEE.RaiseEvent(args);
-                }
-            }
-        }
-
-        public int[] cats;
-        int bmaxxp = 0;
-
-        public Skill(ModEntry LE, string name, int xp, double? xp_mod = null, List<int> xp_table = null, int[] cats = null)
-        {
-            if (xp_table != null && xp_table.Count > 0)
-            {
-                bmaxxp = xp_table.Max();
-            }
-
-            this.LE = LE;
-            this.name = name;
-            this.key = LE.skills.Count;
-            args = new EXPEventArgs();
-            args.key = key;
-            this.xp_table = xp_table ?? new List<int>();
-            if (xp_mod == null)
-                this.xp_mod = 1.0;
-            else
-                this.xp_mod = xp_mod.Value;
-
-            this.cats = cats ?? new int[0];
-
-            if (key == 0) Level = Game1.player.farmingLevel.Value;
-            else if (key == 1) Level = Game1.player.fishingLevel.Value;
-            else if (key == 2) Level = Game1.player.foragingLevel.Value;
-            else if (key == 3) Level = Game1.player.miningLevel.Value;
-            else if (key == 4) Level = Game1.player.combatLevel.Value;
-            else Level = 0;
-
-            generateTable(101);
-
-            if (key < 5)
-            {
-                Game1.player.experiencePoints[key] = xp;
-                XP = xp;
-            }
-            else
-                XP = xp;
-
-            checkForLevelUp();
-        }
-        public void checkForLevelUp()
-        {
-            int l = getLevByXP();
-            if (l != level)
-            {
-                lvlbyxp = true;
-                level = l;
-            }
-        }
-
-        public int getReqXP(int lev)
-        {
-            if (lev < 0) return 0; // Guard against negative index
-            if (xp_table.Count > lev)
-                return xp_table[lev];
-            else
-                generateTable(lev);
-            return xp_table[lev];
-        }
-
-        public void generateTable(int lev)
-        {
-            for (int i = xp_table.Count; i <= lev; i++)
-            {
-                int exp = getXPByLev(i);
-                xp_table.Add(exp);
-            }
-        }
-
-        public int getXPByLev(int i)
-        {
-            if (i <= 0) return 0; // Base case
-            if (xp_table.Count > i)
-                return xp_table[i];
-
-            if (i < 45)
-                return getXPByLev(i - 1) + 300 + (int)Math.Round(1000 * i * xp_mod);
-            else
-                return getXPByLev(i - 1) + 300 + (int)Math.Round(((i * i * i * 0.5)) * xp_mod);
-        }
-
-        public int getLevByXP()
-        {
-            int l = 0;
-            if (xp_table.Count > 0 && xp > xp_table.Max())
-                l = xp_table.Count - 1;
-
-            for (int i = 0; i < xp_table.Count; i++)
-            {
-                if (xp <= xp_table[i])
-                {
-                    l = i;
-                    break;
-                }
-            }
-            return l;
         }
     }
 }
